requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.0
constantly==15.1.0
cryptography==40.0.2
dateparser==1.1.8
defusedxml==0.8.0rc2
Django==4.2
django-appconf==1.0.5
django-templated-mail==1.1.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
gunicorn==20.0.4
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
oauthlib==3.2.2
Pillow==9.5.0
psycopg2-binary==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
redis==4.5.4
regex==2023.3.23
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sorl-thumbnail==12.9.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.6
zope.interface==6.0
